+++ fxn/libs/linux/__init__.py
@@ -1,4 +1,0 @@
-# 
-#   Function
-#   Copyright © 2024 NatML Inc. All Rights Reserved.
-#