+++ fxn/types/__init__.py
@@ -6,8 +6,7 @@
 from .dtype import Dtype
 from .environment import EnvironmentVariable
 from .prediction import Prediction, PredictionResource
-from .predictor import Acceleration, AccessMode, EnumerationMember, Parameter, Predictor, PredictorStatus, PredictorType, Signature
+from .predictor import Acceleration, AccessMode, EnumerationMember, Parameter, Predictor, PredictorStatus, Signature
 from .profile import Profile
 from .storage import UploadType
 from .user import User
-from .value import Value