+++ fxn/types/predictor.py
@@ -3,21 +3,24 @@
 #   Copyright © 2024 NatML Inc. All Rights Reserved.
 #
 
-from enum import Enum
-from pydantic import AliasChoices, BaseModel, Field
-from typing import List, Optional, Tuple, Union
+from enum import Enum, IntFlag
+from io import BytesIO
+from numpy.typing import NDArray
+from PIL import Image
+from pydantic import AliasChoices, BaseModel, ConfigDict, Field
+from typing import Any, Dict, List, Optional, Tuple
 
 from .dtype import Dtype
 from .profile import Profile
-from .value import Value
 
-class Acceleration (str, Enum):
+class Acceleration (IntFlag):
     """
     Predictor acceleration.
     """
-    CPU = "CPU"
-    A40 = "A40"
-    A100 = "A100"
+    Default = 0,
+    CPU = 1 << 0,
+    GPU = 1 << 1,
+    NPU = 1 << 2
 
 class AccessMode (str, Enum):
     """
@@ -25,13 +28,6 @@ class AccessMode (str, Enum):
     """
     Public = "PUBLIC"
     Private = "PRIVATE"
-
-class PredictorType (str, Enum):
-    """
-    Predictor type.
-    """
-    Cloud = "CLOUD"
-    Edge = "EDGE"
 
 class PredictorStatus (str, Enum):
     """
@@ -51,7 +47,7 @@ class EnumerationMember (BaseModel):
         value (str | int): Enumeration member value.
     """
     name: str = Field(description="Enumeration member name.")
-    value: Union[str, int] = Field(description="Enumeration member value.")
+    value: str | int = Field(description="Enumeration member value.")
 
 class Parameter (BaseModel):
     """
@@ -64,7 +60,7 @@ class Parameter (BaseModel):
         optional (bool): Whether the parameter is optional.
         range (tuple): Parameter value range for numeric parameters.
         enumeration (list): Parameter value choices for enumeration parameters.
-        default_value (Value): Parameter default value.
+        default_value (str | float | int | bool | ndarray | list | dict | PIL.Image | BytesIO): Parameter default value.
         value_schema (dict): Parameter JSON schema. This is only populated for `list` and `dict` parameters.
     """
     name: str = Field(description="Parameter name.")
@@ -73,8 +69,9 @@ class Parameter (BaseModel):
     optional: Optional[bool] = Field(default=None, description="Whether the parameter is optional.")
     range: Optional[Tuple[float, float]] = Field(default=None, description="Parameter value range for numeric parameters.")
     enumeration: Optional[List[EnumerationMember]] = Field(default=None, description="Parameter value choices for enumeration parameters.")
-    default_value: Optional[Value] = Field(default=None, description="Parameter default value.", serialization_alias="defaultValue", validation_alias=AliasChoices("default_value", "defaultValue"))
+    default_value: Optional[str | float | int | bool | NDArray | List[Any] | Dict[str, Any] | Image.Image | BytesIO] = Field(default=None, description="Parameter default value.", serialization_alias="defaultValue", validation_alias=AliasChoices("default_value", "defaultValue"))
     value_schema: Optional[dict] = Field(default=None, description="Parameter JSON schema. This is only populated for `list` and `dict` parameters.", serialization_alias="schema", validation_alias=AliasChoices("schema", "value_schema"))
+    model_config = ConfigDict(arbitrary_types_allowed=True)
 
 class Signature (BaseModel):
     """
@@ -95,7 +92,6 @@ class Predictor (BaseModel):
         tag (str): Predictor tag.
         owner (Profile): Predictor owner.
         name (str): Predictor name.
-        type (PredictorType): Predictor type.
         status (PredictorStatus): Predictor status.
         access (AccessMode): Predictor access.
         signature (Signature): Predictor signature.
@@ -103,13 +99,11 @@ class Predictor (BaseModel):
         description (str): Predictor description.
         card (str): Predictor card.
         media (str): Predictor media URL.
-        acceleration (Acceleration): Predictor acceleration. This only applies to cloud predictors.
         license (str): Predictor license URL.
     """
     tag: str = Field(description="Predictor tag.")
     owner: Profile = Field(description="Predictor owner.")
     name: str = Field(description="Predictor name.")
-    type: PredictorType = Field(description="Predictor type.")
     status: PredictorStatus = Field(description="Predictor status.")
     access: AccessMode = Field(description="Predictor access.")
     signature: Signature = Field(description="Predictor signature.")
@@ -117,5 +111,4 @@ class Predictor (BaseModel):
     description: Optional[str] = Field(default=None, description="Predictor description.")
     card: Optional[str] = Field(default=None, description="Predictor card.")
     media: Optional[str] = Field(default=None, description="Predictor media URL.")
-    acceleration: Optional[Acceleration] = Field(default=None, description="Predictor acceleration. This only applies to cloud predictors.")
     license: Optional[str] = Field(default=None, description="Predictor license URL.")